:recycle: [#6275] Remove is_visible_in_frontend
Basically, it did exactly the same as `is_hidden`, but without the use of the component registry.

To avoid having to solve bugs in multiple places at the same time, I have removed it completely.

Note that `is_hidden` needs a slight update to deal with unexpected components.
